moved header, did the enhanced demo Dan showed
