+++ src/App.css
@@ -1,14 +1,28 @@
-* {
+body {
   background-color: darkkhaki;
+}
+
+.header {
+  width: 100%;
+  background-color: #282c34;
+  color: red;
+}
+
+.footer {
+  width: 100%;
+  background-color: #282c34;
+  color: red;
+  position: absolute;
+  bottom: 0;
 }
 
 .App {
   text-align: center;
 }
-
+/* 
 .App-header {
   background-color: #282c34;
-  /* min-height: 100vh; */
+  min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -25,4 +39,4 @@
   justify-content: center;
   font-size: calc(10px + 2vmin);
   color: white;
-}
+} */
