+++ src/App.js
@@ -1,21 +1,21 @@
+import { ListProvider } from './ Context/ListContext';
 import './App.css';
+import Header from './Components/Header/Header';
 import Home from './Views/Home';
+import Footer from './Components/Footer/Footer';
 // import AddItem from './Components/Add/AddItem';
 // import ItemList from './Components/List/ItemList';
 // import EditItem from './Components/Edit/EditItems';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">app header</header>
-      {/* <h1>Shopping List</h1> */}
-      <Home />
-      {/* <AddItem />
-      dont need add item, just want to see it on page atm
-      <ItemList />
-      <EditItem /> */}
-      <footer className="App-footer">app footer</footer>
-    </div>
+    <ListProvider>
+      <div className="App">
+        <Header />
+        <Home />
+        <Footer />
+      </div>
+    </ListProvider>
   );
 }
