+++ src/Components/Footer/Footer.js
@@ -1,0 +1,9 @@
+import { useList } from '../../ Context/ListContext';
+
+export default function Footer() {
+  return (
+    <footer className="footer">
+      <div>this is the footer</div>
+    </footer>
+  );
+}
