+++ src/Components/Header/Header.js
@@ -1,0 +1,9 @@
+import { useList } from '../../ Context/ListContext';
+
+export default function Header() {
+  return (
+    <header className="header">
+      <div>add counter to this header</div>
+    </header>
+  );
+}
